Font updated

+++ src/graphics/glyphs.rs
@@ -6,12 +6,12 @@ use super::{
 #[derive(Copy, Clone, Debug)]
 pub struct Glyph {
     pub size: GlyphSize,
-    pub delta_x: f32,
+    pub delta_x: i32,
     pub code: u32,
 }
 
 impl Glyph {
-    pub fn new(size: GlyphSize, delta_x: f32, code: u32) -> Self { Glyph { size, delta_x, code } }
+    pub fn new(size: GlyphSize, delta_x: i32, code: u32) -> Self { Glyph { size, delta_x, code } }
 }
 
 #[derive(Debug)]
